+++ lib/src/io/ifstream.cc
@@ -14,17 +14,17 @@ bool InputFileStreamBuffer::NextPage() {
     auto data = buffer_.GetData();
     auto data_offset = page_id << filesystem_buffer_->GetPageSizeShift();
     assert(data_offset < data_end_);
-    auto data_size = std::min<size_t>(data.size(), data_end_ - data_offset);
+    auto data_size = std::min<uint64_t>(data.size(), data_end_ - data_offset);
     setg(data.data(), data.data(), data.data() + data_size);
     return true;
 }
 
 std::streamsize InputFileStreamBuffer::xsgetn(char* out, std::streamsize want) {
     auto base = out;
-    auto left = std::min<size_t>(want, file_.GetSize() - GetPosition());
+    auto left = std::min<uint64_t>(want, file_.GetSize() - GetPosition());
     assert((egptr() - gptr()) <= (file_.GetSize() - GetPosition()));
     while (left > 0 && (gptr() < egptr() || NextPage())) {
-        auto m = std::min<size_t>(egptr() - gptr(), left);
+        auto m = std::min<uint64_t>(egptr() - gptr(), left);
         std::memcpy(out, gptr(), m);
         gbump(m);
         out += m;
@@ -35,13 +35,13 @@ std::streamsize InputFileStreamBuffer::xsgetn(char* out, std::streamsize want) {
 
 InputFileStreamBuffer::pos_type InputFileStreamBuffer::seekoff(off_type n, std::ios_base::seekdir dir,
                                                                std::ios_base::openmode) {
-    size_t pos;
+    uint64_t pos;
     if (dir == std::ios_base::beg) {
         pos = n;
     } else if (dir == std::ios_base::end) {
         pos = file_.GetSize() - n;
     } else {
-        pos = std::min<size_t>(file_.GetSize(), pos + n);
+        pos = std::min<uint64_t>(file_.GetSize(), pos + n);
     }
     auto page_id = pos >> filesystem_buffer_->GetPageSizeShift();
     auto page_ofs = pos - (page_id << filesystem_buffer_->GetPageSizeShift());
@@ -60,7 +60,7 @@ InputFileStreamBuffer::pos_type InputFileStreamBuffer::seekpos(pos_type p, std::
     return p;
 }
 
-void InputFileStreamBuffer::Slice(size_t offset, size_t size) {
+void InputFileStreamBuffer::Slice(uint64_t offset, uint64_t size) {
     // Determine range
     auto file_size = file_.GetSize();
     auto begin = std::min(file_size, offset);
